Fixed issues with some core matrix code

+++ src/utils/matrix.h
@@ -7,11 +7,11 @@ public:
     static bool CreateBackup(RwFrame* frame);
     static void RestoreBackup(RwMatrix* dest, RwMatrix* backup);
 
-    static float GetRotationX(RwMatrix *matrix);
-    static float GetRotationY(RwMatrix *matrix);
-    static float GetRotationZ(RwMatrix *matrix);
+    static double GetRotationX(RwMatrix *matrix);
+    static double GetRotationY(RwMatrix *matrix);
+    static double GetRotationZ(RwMatrix *matrix);
     static void ResetRotation(RwMatrix *matrix);
-    static void SetRotationX(RwMatrix *matrix, float angle);
-    static void SetRotationY(RwMatrix *matrix, float angle);
-    static void SetRotationZ(RwMatrix *matrix, float angle);
+    static void SetRotationX(RwMatrix *matrix, double angle);
+    static void SetRotationY(RwMatrix *matrix, double angle);
+    static void SetRotationZ(RwMatrix *matrix, double angle);
 };
